voice-agent/pyproject.toml
[project]
name = "voice_agent_project"
version = "0.1.0"
description = "Simple voice agent using Pipecat"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[daily,silero,webrtc]",
    "python-dotenv",
    "openai",
    "deepgram-sdk>=3.2.0,<4.0.0",
    "cartesia",
    "aiohttp",
    "wikipedia>=1.4.0",
    "sounddevice>=0.5.3",
    "pyaudio>=0.2.14",
    "fastapi>=0.121.2",
    "uvicorn>=0.38.0",
    "pipecat-ai-small-webrtc-prebuilt>=1.0.0",
    "onnxruntime>=1.23.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["wikipedia_voice_bot.py"]

